Fix scheduling for PowerAuth auditing
