+++ powerauth-java-server/src/main/java/io/getlime/security/powerauth/app/server/configuration/WebServiceConfig.java
@@ -18,8 +18,6 @@
 
 package io.getlime.security.powerauth.app.server.configuration;
 
-import com.wultra.core.audit.base.Audit;
-import com.wultra.core.audit.base.AuditFactory;
 import io.getlime.security.powerauth.app.server.service.exceptions.ActivationRecoveryException;
 import io.getlime.security.powerauth.app.server.service.exceptions.GenericServiceException;
 import io.getlime.security.powerauth.app.server.service.exceptions.SoapFaultExceptionResolver;
@@ -28,7 +26,6 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -56,14 +53,11 @@ import java.util.Properties;
  */
 @EnableWs
 @Configuration
-@ComponentScan(basePackages = {"com.wultra.core.audit.base"})
 public class WebServiceConfig extends WsConfigurerAdapter {
 
     private final UserDetailsService userDetailsService;
 
     private PowerAuthServiceConfiguration configuration;
-
-    private final AuditFactory auditFactory;
 
     /**
      * Setter for configuration injection.
@@ -77,21 +71,10 @@ public class WebServiceConfig extends WsConfigurerAdapter {
     /**
      * Constructor that accepts an instance of UserDetailsService for autowiring.
      * @param userDetailsService UserDetailsService instance.
-     * @param auditFactory Audit factory.
      */
     @Autowired
-    public WebServiceConfig(@Qualifier("integrationUserDetailsService") UserDetailsService userDetailsService, AuditFactory auditFactory) {
+    public WebServiceConfig(@Qualifier("integrationUserDetailsService") UserDetailsService userDetailsService) {
         this.userDetailsService = userDetailsService;
-        this.auditFactory = auditFactory;
-    }
-
-    /**
-     * Bean instance of audit class produced using a factory.
-     * @return Audit instance.
-     */
-    @Bean
-    public Audit audit() {
-        return auditFactory.getAudit();
     }
 
     /**
